created and styled footer.php
Using php, we only have to create the footer in one file, then we can include it in as many files we want across our site. Because the footer will be on almost all of our pages, this minimizes how much html we have to write in the future. (This was a very basic implementation, I expect the style and possibly the html to evolve throughout our project)