created ajax loading of  header from data.json information as well as styled the header for all pages of the site
I created a after sudo element on every even a tag in the list to give some sepparation to the links. I did this with a local css variable, so we can change the height if we end up wanting it different. (if we change the site font it is possible this will not look as good)