I was able to make the background fit the whole page of the website.
I used inspect to see how I could make the background cover the entire page. That is why I used body,html as a selector to make sure that the background fit the entire page. I also made a login.css so that all of the css style applies for the login page.