I was able to make the background cover the entire page.
I wanted to include a purple cloud background, so that it balanced with the purple color theme background. I also used inspect to help me see how I can have the background image cover the entire page. I used a font from google fonts to help with the font on the field area.